// packages/providers/src/live/az-schema.ts
export interface components {
    schemas: {
// ...
        LoginResponse: {
            /** @description The JWT token */
            jwt?: string;
            /** @description Is ownerAgent */
            ownerAgent?: boolean;
        };
// ...
        Employee: {
            /** @description The employee ID */
            id?: number;
            /** @description First name */
            firstname?: string;
            /** @description Last name */
            lastname?: string;
            /** @description Email */
            email?: string;
            /** @description Phone */
            phone?: string;
            /** @description Is the employee a producer */
            isProducer?: boolean;
            /** @description is the employee active */
            isActive?: boolean;
            /** @description Is the employee owner agent */
            isOwner?: boolean;
            /** @description The user ID */
            userId?: number;
            /** @description The avatar url */
            avatar?: string;
            /** @description External system name from AMS integration settings */
            externalSystem?: string;
            /** @description Array of external IDs */
            externalIds?: string[];
        };
// ...
        LeadSearchResponse: components["schemas"]["BaseSearchResponse"] & {
            /** @description A list of leads */
            leads?: components["schemas"]["Lead"][];
            /** @description Quote amount */
            quoteAmount?: number;
        };
// ...
        Quote: components["schemas"]["Opportunity"] & {
            /** @description whether the quote is sold */
            sold?: boolean;
            /** @description The carrier name */
            carrierName?: string;
            /** @description The policy type */
            productName?: string;
            /** @description The potential revenue */
            potentialRevenue?: number;
            /** @description The effective date */
            effectiveDate?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}